improved structure with wizard tools as global file in the wizard context
