+++ QgisModelBaker/gui/workflow_wizard/import_data_configuration_page.py
@@ -57,87 +57,11 @@ from QgisModelBaker.gui.panel.log_panel import LogPanel
 from QgisModelBaker.gui.dataset_manager import DatasetManagerDialog
 from QgisModelBaker.gui.ili2db_options import Ili2dbOptionsDialog
 
+import QgisModelBaker.gui.workflow_wizard.wizard_tools as wizard_tools
+
 from ...utils import get_ui_class
 
 PAGE_UI = get_ui_class('workflow_wizard/import_data_configuration.ui')
-
-DEFAULT_DATASETNAME = "Baseset"
-class SourceModel(QStandardItemModel):
-
-    print_info = pyqtSignal([str], [str, str])
-    class Roles(Enum):
-        NAME = Qt.UserRole + 1
-        TYPE = Qt.UserRole + 2
-        PATH = Qt.UserRole + 3
-        DATASET_NAME = Qt.UserRole + 5
-
-        def __int__(self):
-            return self.value
-
-    def __init__(self):
-        super().__init__()
-        self.setColumnCount(2)
-
-    def flags(self, index):
-        if index.column() > 0:
-            return Qt.ItemIsSelectable | Qt.ItemIsEnabled | Qt.ItemIsEditable
-        return Qt.ItemIsSelectable | Qt.ItemIsEnabled
-
-    def headerData(self, section, orientation, role):
-        if orientation == Qt.Vertical and role == Qt.DisplayRole:
-            return "↑ ↓"
-        return QStandardItemModel.headerData(self, section,orientation,role)
-
-    def data(self, index, role):
-        item = self.item(index.row(), index.column())
-        if role == Qt.DisplayRole:
-            if index.column() > 0:
-                return item.data(int(SourceModel.Roles.DATASET_NAME))
-            if item.data(int(SourceModel.Roles.TYPE)) != 'model':
-                return self.tr('{} ({})').format(item.data(int(Qt.DisplayRole)), item.data(int(SourceModel.Roles.PATH)))
-        if role == Qt.DecorationRole:
-            if index.column() == 0:
-                type = 'data'
-                if item.data(int(SourceModel.Roles.TYPE)) and item.data(int(SourceModel.Roles.TYPE)).lower() in ['model','ili', 'xtf', 'xml']:
-                    type = item.data(int(SourceModel.Roles.TYPE)).lower()
-                return QIcon(os.path.join(os.path.dirname(__file__), f'../../images/file_types/{type}.png'))
-        return item.data(int(role))
-
-    def add_source(self, name, type, path):
-        if self.source_in_model(name, type, path):
-            self.print_info.emit(self.tr("Source alread added {} ({})").format(
-                name, path if path else 'repository'))
-            return
-
-        item = QStandardItem()
-        item.setData(name, int(Qt.DisplayRole))
-        item.setData(name, int(SourceModel.Roles.NAME))
-        item.setData(type, int(SourceModel.Roles.TYPE))
-        item.setData(path, int(SourceModel.Roles.PATH))
-        self.appendRow([item, QStandardItem()])
-
-        self.print_info.emit(self.tr("Add source {} ({})").format(
-            name, path if path else 'repository'))
-
-    def source_in_model(self, name, type, path):
-        match_existing = self.match(self.index(
-            0, 0), SourceModel.Roles.NAME, name, -1, Qt.MatchExactly)
-        if match_existing and type == match_existing[0].data(int(SourceModel.Roles.TYPE)) and path == match_existing[0].data(int(SourceModel.Roles.PATH)):
-            return True
-        return False
-
-    def setData(self, index, data, role):
-        if index.column() > 0:
-            return QStandardItemModel.setData(self, index, data, int(SourceModel.Roles.DATASET_NAME))
-        return QStandardItemModel.setData(self, index, data, role)
-
-    def remove_sources(self, indices):
-        for index in sorted(indices):
-            path = index.data(int(SourceModel.Roles.PATH))
-            self.print_info.emit(self.tr("Remove source {} ({})").format(
-                index.data(int(SourceModel.Roles.NAME)), path if path else 'repository'))
-            self.removeRow(index.row())
-
 class DatasetComboDelegate(QStyledItemDelegate):
     def __init__(self, parent, db_connector):
         super().__init__(parent)
@@ -150,12 +74,12 @@ class DatasetComboDelegate(QStyledItemDelegate):
         self.editor.addItems(self.items)
         return self.editor
     def setEditorData(self, editor, index):
-        value = index.data(int(SourceModel.Roles.DATASET_NAME))
+        value = index.data(int(wizard_tools.SourceModel.Roles.DATASET_NAME))
         num = self.items.index(value) if value in self.items else 0
         editor.setCurrentIndex(num)
     def setModelData(self, editor, model, index):
         value = editor.currentText()
-        model.setData(index, value, int(SourceModel.Roles.DATASET_NAME))
+        model.setData(index, value, int(wizard_tools.SourceModel.Roles.DATASET_NAME))
     def updateEditorGeometry(self, editor, option, index):
         editor.setGeometry(option.rect)
 class ImportDataConfigurationPage(QWizardPage, PAGE_UI):
@@ -202,9 +126,9 @@ class ImportDataConfigurationPage(QWizardPage, PAGE_UI):
             # set defaults
             for row in range( self.workflow_wizard.import_data_file_model.rowCount() ):
                 index = self.workflow_wizard.import_data_file_model.index(row, 1)
-                value = index.data(int(SourceModel.Roles.DATASET_NAME))
+                value = index.data(int(wizard_tools.SourceModel.Roles.DATASET_NAME))
                 if not value:
-                    self.workflow_wizard.import_data_file_model.setData(index, DEFAULT_DATASETNAME, int(SourceModel.Roles.DATASET_NAME))
+                    self.workflow_wizard.import_data_file_model.setData(index, wizard_tools.DEFAULT_DATASETNAME, int(wizard_tools.SourceModel.Roles.DATASET_NAME))
                 self.file_table_view.setItemDelegateForColumn(1, DatasetComboDelegate(self, self.db_connector))
         else:
             self.file_table_view.setColumnHidden(1, True)
