+++ QgisModelBaker/gui/workflow_wizard/workflow_wizard.py
@@ -18,23 +18,10 @@
  ***************************************************************************/
 """
 
-from enum import Enum
-import os
-import re
-
 from qgis.PyQt.QtCore import (
-    QSortFilterProxyModel,
     QTimer,
     Qt,
     QEventLoop,
-    pyqtSignal,
-    QStringListModel
-)
-
-from qgis.PyQt.QtGui import (
-    QStandardItemModel,
-    QStandardItem,
-    QIcon
 )
 
 from qgis.PyQt.QtWidgets import (
@@ -43,27 +30,20 @@ from qgis.PyQt.QtWidgets import (
     QVBoxLayout
 )
 
-from qgis.gui import (
-    QgsMessageBar
+from QgisModelBaker.gui.workflow_wizard.wizard_tools import(
+    SourceModel,
+    ImportDataModel,
+    ImportModelsModel,
+    ExportDatasetsModel,
+    ExportModelsModel,
+    TransferExtensions,
 )
 
-from QgisModelBaker.gui.workflow_wizard.intro_page import IntroPage
-from QgisModelBaker.gui.workflow_wizard.import_source_selection_page import ImportSourceSeletionPage
-from QgisModelBaker.gui.workflow_wizard.database_selection_page import DatabaseSelectionPage
-from QgisModelBaker.gui.workflow_wizard.import_schema_configuration_page import ImportSchemaConfigurationPage
-from QgisModelBaker.gui.workflow_wizard.execution_page import ExecutionPage
-from QgisModelBaker.gui.workflow_wizard.project_creation_page import ProjectCreationPage
-from QgisModelBaker.gui.workflow_wizard.import_data_configuration_page import ImportDataConfigurationPage
-from QgisModelBaker.gui.workflow_wizard.export_data_configuration_page import ExportDataConfigurationPage
-from QgisModelBaker.gui.workflow_wizard.import_data_configuration_page import SourceModel
 from QgisModelBaker.gui.panel.log_panel import LogPanel
 
-from QgisModelBaker.libili2db.globals import DbIliMode, displayDbIliMode, DbActionType
+from QgisModelBaker.libili2db.globals import DbActionType
 
-from QgisModelBaker.libili2db.ilicache import (
-    IliCache,
-    IliToppingFileCache
-)
+from QgisModelBaker.libili2db.ilicache import IliToppingFileCache
 
 from QgisModelBaker.libili2db.ili2dbconfig import (
     UpdateDataConfiguration,
@@ -74,256 +54,14 @@ from QgisModelBaker.libili2db.ili2dbconfig import (
 from ...libqgsprojectgen.db_factory.db_simple_factory import DbSimpleFactory
 from ...libqgsprojectgen.dbconnector.db_connector import DBConnectorError
 
-IliExtensions = ['ili']
-TransferExtensions = ['xtf', 'XTF', 'itf', 'ITF', 'pdf',
-                      'PDF', 'xml', 'XML', 'xls', 'XLS', 'xlsx', 'XLSX']
-class ImportModelsModel(SourceModel):
-
-    def __init__(self):
-        super().__init__()
-        self._checked_models = {}
-
-    def refresh_model(self, filtered_source_model, db_connector=None, silent=False):
-
-        self.clear()
-        previously_checked_models = self._checked_models
-        self._checked_models = {}
-
-        # models from db
-        db_modelnames = self.db_modelnames(db_connector)
-
-        # models from the repos
-        models_from_repo = []
-        filtered_source_model.setFilterFixedString('model')
-        for r in range(0, filtered_source_model.rowCount()):
-            filtered_source_model_index = filtered_source_model.index(r, 0)
-            modelname = filtered_source_model_index.data(
-                int(SourceModel.Roles.NAME))
-            if modelname:
-                enabled = modelname not in db_modelnames
-                self.add_source(modelname, filtered_source_model_index.data(int(SourceModel.Roles.TYPE)), filtered_source_model_index.data(
-                    int(SourceModel.Roles.PATH)), previously_checked_models.get(modelname, Qt.Checked) if enabled else Qt.Unchecked, enabled)
-                models_from_repo.append(
-                    filtered_source_model_index.data(int(SourceModel.Roles.NAME)))
-                if not silent:
-                    self.print_info.emit(
-                        self.tr("- Append (repository) model {}{}").format(modelname, " (inactive because it already exists in the database)" if not enabled else ''))
-
-        # models from the files
-        models_from_ili_files = []
-        filtered_source_model.setFilterFixedString('ili')
-        for r in range(0, filtered_source_model.rowCount()):
-            filtered_source_model_index = filtered_source_model.index(r, 0)
-            ili_file_path = filtered_source_model_index.data(
-                int(SourceModel.Roles.PATH))
-            self.ilicache = IliCache(None, ili_file_path)
-            models = self.ilicache.process_ili_file(ili_file_path)
-            for model in models:
-                if model['name']:
-                    enabled = model['name'] not in db_modelnames
-                    self.add_source(model['name'], filtered_source_model_index.data(int(SourceModel.Roles.TYPE)), filtered_source_model_index.data(
-                        int(SourceModel.Roles.PATH)), previously_checked_models.get(model['name'], Qt.Checked if model is models[-1] and enabled else Qt.Unchecked), enabled)
-                    models_from_ili_files.append(model['name'])
-                    if not silent:
-                        self.print_info.emit(
-                            self.tr("- Append (file) model {}{} from {}").format(model['name'], " (inactive because it already exists in the database)" if not enabled else '', ili_file_path))
-
-        # models from the transfer files (not yet implemented)
-        filtered_source_model.setFilterRegExp('|'.join(TransferExtensions))
-        for r in range(0, filtered_source_model.rowCount()):
-            index = filtered_source_model.index(r, 0)
-            xtf_file_path = index.data(int(SourceModel.Roles.PATH))
-            if not silent:
-                self.print_info.emit(
-                    self.tr("Get models from the transfer file ({}) is not yet implemented").format(xtf_file_path))
-
-        return self.rowCount()
-
-    def db_modelnames(self, db_connector=None):
-        modelnames = list()
-        if db_connector:
-            if db_connector.db_or_schema_exists() and db_connector.metadata_exists():
-                db_models = db_connector.get_models()
-                for db_model in db_models:
-                    regex = re.compile(r'(?:\{[^\}]*\}|\s)')
-                    for modelname in regex.split(db_model['modelname']):
-                        modelnames.append(modelname.strip())
-        return modelnames
-
-    def add_source(self, name, type, path, checked, enabled):
-        item = QStandardItem()
-        self._checked_models[name] = checked
-        item.setFlags(Qt.ItemIsSelectable |
-                      Qt.ItemIsEnabled if enabled else Qt.NoItemFlags)
-        item.setData(name, int(Qt.DisplayRole))
-        item.setData(name, int(SourceModel.Roles.NAME))
-        item.setData(type, int(SourceModel.Roles.TYPE))
-        item.setData(path, int(SourceModel.Roles.PATH))
-        self.appendRow(item)
-
-    def data(self, index, role):
-        if role == Qt.DisplayRole:
-            return self.tr("{}{}").format(SourceModel.data(self, index, (Qt.DisplayRole)), "" if index.flags() & Qt.ItemIsEnabled else " (already in the database)")
-        if role == Qt.CheckStateRole:
-            return self._checked_models[self.data(index, int(SourceModel.Roles.NAME))]
-        return SourceModel.data(self, index, role)
-
-    def setData(self, index, role, data):
-        if role == Qt.CheckStateRole:
-            self.beginResetModel()
-            self._checked_models[self.data(
-                index, int(SourceModel.Roles.NAME))] = data
-            self.endResetModel()
-
-    def flags(self, index):
-        item = self.item(index.row(), index.column())
-        if item:
-            return item.flags()
-        return Qt.NoItemFlags
-
-    def check(self, index):
-        if index.flags() & Qt.ItemIsEnabled:
-            if self.data(index, Qt.CheckStateRole) == Qt.Checked:
-                self.setData(index, Qt.CheckStateRole, Qt.Unchecked)
-            else:
-                self.setData(index, Qt.CheckStateRole, Qt.Checked)
-
-    def import_sessions(self):
-        sessions = {}
-        for r in range(0, self.rowCount()):
-            item = self.index(r, 0)
-            if item.data(int(Qt.Checked)):
-                type = item.data(int(SourceModel.Roles.TYPE))
-                model = item.data(int(SourceModel.Roles.NAME))
-                source = item.data(int(SourceModel.Roles.PATH)
-                                   ) if type != 'model' else 'repository '+model
-
-                if self._checked_models[model] == Qt.Checked:
-                    models = []
-                    if source in sessions:
-                        models = sessions[source]['models']
-                    else:
-                        sessions[source] = {}
-                    models.append(model)
-                    sessions[source]['models'] = models
-        return sessions
-
-    def checked_models(self):
-        return [model for model in self._checked_models.keys() if self._checked_models[model] == Qt.Checked]
-
-class ImportDataModel(QSortFilterProxyModel):
-
-    print_info = pyqtSignal([str], [str, str])
-
-    def __init__(self):
-        super().__init__()
-        self._checked_models = {}
-
-    def import_sessions(self, order_list):
-        sessions = {}
-        i = 0
-        for r in order_list:
-            source = self.index(r, 0).data(int(SourceModel.Roles.PATH))
-            dataset = self.index(r, 1).data(
-                int(SourceModel.Roles.DATASET_NAME))
-            sessions[source] = {}
-            sessions[source]['dataset'] = dataset
-            i += 1
-        return sessions
-        
-class ExportModelsModel(QStringListModel):
-    # at dave: can we find synergies here with the dataset model of the dataset manager
-    # dann wärs ein standarditemmodel - das handling der checked_models (bzw. checked_datasets wär dann wie im ImportModelsModel)
-    # evtl. könnte man es dann auch gleich fürs ExportModelsModel machen
-    # oder zumindest ExportDatasetsModel und ExportModelsModel
-
-    blacklist = ['CHBaseEx_MapCatalogue_V1', 'CHBaseEx_WaterNet_V1', 'CHBaseEx_Sewage_V1', 'CHAdminCodes_V1',
-                    'AdministrativeUnits_V1', 'AdministrativeUnitsCH_V1', 'WithOneState_V1',
-                    'WithLatestModification_V1', 'WithModificationObjects_V1', 'GraphicCHLV03_V1', 'GraphicCHLV95_V1',
-                    'NonVector_Base_V2', 'NonVector_Base_V3', 'NonVector_Base_LV03_V3_1', 'NonVector_Base_LV95_V3_1',
-                    'GeometryCHLV03_V1', 'GeometryCHLV95_V1', 'InternationalCodes_V1', 'Localisation_V1',
-                    'LocalisationCH_V1', 'Dictionaries_V1', 'DictionariesCH_V1', 'CatalogueObjects_V1',
-                    'CatalogueObjectTrees_V1', 'AbstractSymbology', 'CodeISO', 'CoordSys', 'GM03_2_1Comprehensive',
-                    'GM03_2_1Core', 'GM03_2Comprehensive', 'GM03_2Core', 'GM03Comprehensive', 'GM03Core',
-                    'IliRepository09', 'IliSite09', 'IlisMeta07', 'IliVErrors', 'INTERLIS_ext', 'RoadsExdm2ben',
-                    'RoadsExdm2ben_10', 'RoadsExgm2ien', 'RoadsExgm2ien_10', 'StandardSymbology', 'StandardSymbology',
-                    'Time', 'Units']
-
-    def __init__(self):
-        super().__init__()
-        self._checked_models = None
-
-    def refresh_model(self, db_connector=None):
-        modelnames = []
-        
-        if db_connector:
-            if db_connector.db_or_schema_exists() and db_connector.metadata_exists():
-                db_models = db_connector.get_models()
-                for db_model in db_models:
-                    regex = re.compile(r'(?:\{[^\}]*\}|\s)')
-                    for modelname in regex.split(db_model['modelname']):
-                        if modelname and modelname not in ExportModelsModel.blacklist:
-                            modelnames.append(modelname.strip())
-
-        self.setStringList(modelnames)
-
-        self._checked_models = {modelname: Qt.Checked for modelname in modelnames}
-
-        return self.rowCount()
-
-    def data(self, index, role):
-        if role == Qt.CheckStateRole:
-            return self._checked_models[self.data(index, Qt.DisplayRole)]
-        else:
-            return QStringListModel.data(self, index, role)
-
-    def setData(self, index, role, data):
-        if role == Qt.CheckStateRole:
-            self._checked_models[self.data(index, Qt.DisplayRole)] = data
-        else:
-            QStringListModel.setData(self, index, role, data)
-
-    def check(self, index):
-        if self.data(index, Qt.CheckStateRole) == Qt.Checked:
-            self.setData(index, Qt.CheckStateRole, Qt.Unchecked)
-        else:
-            self.setData(index, Qt.CheckStateRole, Qt.Checked)
-    
-    def check_all(self):
-        for name in self.stringList():
-            self._checked_models[name] = Qt.Checked
-
-    def checked_models(self):
-        return [modelname for modelname in self.stringList() if self._checked_models[modelname] == Qt.Checked]
-
-        self.editor = QComboBox(parent)
-        self.editor.addItems(self.items)
-class ExportDatasetsModel(QStringListModel):
-    def __init__(self):
-        super().__init__()
-        self._selected_dataset = None
-
-    def refresh_model(self, db_connector=None):
-        datasetnames = []
-        
-        if db_connector and db_connector.db_or_schema_exists():
-            datasets_info = db_connector.get_datasets_info()
-            for record in datasets_info:
-                datasetnames.append(record['datasetname'])
-        self.setStringList(datasetnames)
-        if datasetnames:
-            self._selected_dataset = datasetnames[0]
-        return self.rowCount()
-
-    def flags(self, index):
-        return Qt.ItemIsSelectable | Qt.ItemIsEnabled
-
-    def select(self, row):
-        self._selected_dataset = self.data(self.index(row,0), Qt.DisplayRole)
-    
-    def selected_dataset(self):
-        return self._selected_dataset
-
+from QgisModelBaker.gui.workflow_wizard.intro_page import IntroPage
+from QgisModelBaker.gui.workflow_wizard.import_source_selection_page import ImportSourceSeletionPage
+from QgisModelBaker.gui.workflow_wizard.database_selection_page import DatabaseSelectionPage
+from QgisModelBaker.gui.workflow_wizard.import_schema_configuration_page import ImportSchemaConfigurationPage
+from QgisModelBaker.gui.workflow_wizard.execution_page import ExecutionPage
+from QgisModelBaker.gui.workflow_wizard.project_creation_page import ProjectCreationPage
+from QgisModelBaker.gui.workflow_wizard.import_data_configuration_page import ImportDataConfigurationPage
+from QgisModelBaker.gui.workflow_wizard.export_data_configuration_page import ExportDataConfigurationPage
 class WorkflowWizard (QWizard):
 
     Page_Intro_Id = 1
@@ -351,8 +89,7 @@ class WorkflowWizard (QWizard):
         self.iface = iface
         self.log_panel = parent.log_panel
 
-        # config setup
-        self.db_simple_factory = DbSimpleFactory()
+        # configuration objects are keeped on top level to be able to access them from individual pages
         self.import_schema_configuration = SchemaImportConfiguration()
         self.import_data_configuration = UpdateDataConfiguration()
         self.export_data_configuration = ExportConfiguration()
@@ -360,16 +97,16 @@ class WorkflowWizard (QWizard):
         self.import_data_configuration.base_configuration = base_config
         self.export_data_configuration.base_configuration = base_config
 
-        # models setup
+        # data models are keeped on top level because sometimes they need to be accessed to evaluate the wizard workflow
+        # the source_model keeps all the sources (files or repositories) used and the dataset property
         self.source_model = SourceModel()
         self.source_model.print_info.connect(self.log_panel.print_info)
 
-        self.file_model = QSortFilterProxyModel()
-        self.file_model.setSourceModel(self.source_model)
-        self.file_model.setFilterRole(int(SourceModel.Roles.TYPE))
+        # the import_models_model keeps every single model as entry and a checked state
         self.import_models_model = ImportModelsModel()
         self.import_models_model.print_info.connect(self.log_panel.print_info)
 
+        # the import_data_file_model keeps the filtered out transfer files (from source model) and functions to get ordered import sessions
         self.import_data_file_model = ImportDataModel()
         self.import_data_file_model.print_info.connect(
             self.log_panel.print_info)
@@ -378,16 +115,15 @@ class WorkflowWizard (QWizard):
         self.import_data_file_model.setFilterRegExp(
             '|'.join(TransferExtensions))
         
+        # the export_models_model keeps every single model found in the current database and a checked state
         self.export_models_model = ExportModelsModel()
+        # the export_datasets_models keeps every dataset found in the current database and keeps the selected dataset
         self.export_datasets_model = ExportDatasetsModel()
+        # the current export target is the current set target file for the export. It's keeped top level to have a consequent behavior of those information.
         self.current_export_target = ""
-        # dave oder sollen wir current_export_target aus dem export_data_configuration_page holen?
 
         # pages setup
-        self.intro_page = IntroPage(
-            self, self._current_page_title(self.Page_Intro_Id))
-
-        # import
+        self.intro_page = IntroPage(self, self._current_page_title(self.Page_Intro_Id))
         self.source_seletion_page = ImportSourceSeletionPage(
             self, self._current_page_title(self.Page_ImportSourceSeletion_Id))
         self.import_database_seletion_page = DatabaseSelectionPage(
@@ -402,7 +138,11 @@ class WorkflowWizard (QWizard):
             self, self._current_page_title(self.Page_ImportDataExecution_Id), DbActionType.IMPORT_DATA)
         self.project_creation_page = ProjectCreationPage(
             self, self._current_page_title(self.Page_ProjectCreation_Id))
-
+        self.generate_database_seletion_page = DatabaseSelectionPage(
+            self, self._current_page_title(self.Page_GenerateDatabaseSelection_Id), DbActionType.GENERATE)
+        self.export_database_seletion_page = DatabaseSelectionPage(self, self._current_page_title(self.Page_ExportDatabaseSelection_Id), DbActionType.EXPORT)
+        self.export_data_configuration_page = ExportDataConfigurationPage(self, self._current_page_title(self.Page_ExportDataConfiguration_Id))
+        self.export_data_execution_page = ExecutionPage(self, self._current_page_title(self.Page_ExportDataExecution_Id), DbActionType.EXPORT)
         self.setPage(self.Page_Intro_Id, self.intro_page)
         self.setPage(self.Page_ImportSourceSeletion_Id,
                      self.source_seletion_page)
@@ -417,17 +157,8 @@ class WorkflowWizard (QWizard):
                      self.import_data_execution_page)
         self.setPage(self.Page_ProjectCreation_Id,
                      self.project_creation_page)
-
-        # bake project
-        self.generate_database_seletion_page = DatabaseSelectionPage(
-            self, self._current_page_title(self.Page_GenerateDatabaseSelection_Id), DbActionType.GENERATE)
         self.setPage(self.Page_GenerateDatabaseSelection_Id,
                      self.generate_database_seletion_page)
-
-        # export not yet implemented
-        self.export_database_seletion_page = DatabaseSelectionPage(self, self._current_page_title(self.Page_ExportDatabaseSelection_Id), DbActionType.EXPORT)
-        self.export_data_configuration_page = ExportDataConfigurationPage(self, self._current_page_title(self.Page_ExportDataConfiguration_Id))
-        self.export_data_execution_page = ExecutionPage(self, self._current_page_title(self.Page_ExportDataExecution_Id), DbActionType.EXPORT)
         self.setPage(self.Page_ExportDatabaseSelection_Id,
                      self.export_database_seletion_page)
         self.setPage(self.Page_ExportDataConfiguration_Id,
@@ -444,28 +175,52 @@ class WorkflowWizard (QWizard):
             return self.Page_ImportDatabaseSelection_Id
 
         if self.current_id == self.Page_ImportDatabaseSelection_Id:
-            # update configuration for import data and for import schema and use schema config to save
+            # update configuration for import data 
+            # and for import schema and export data (to have the same db settings for all of them)
+            # and use schema config to save
             self.import_database_seletion_page.update_configuration(
                 self.import_schema_configuration)
             self.import_database_seletion_page.update_configuration(
                 self.import_data_configuration)
+            self.import_database_seletion_page.update_configuration(
+                self.export_data_configuration)
             self.import_database_seletion_page.save_configuration(
                 self.import_schema_configuration)
             if self.refresh_import_models(True):
+                # when there are models to import, we go to the configuration page for schema import
                 return self.Page_ImportSchemaConfiguration_Id
             if self.import_data_file_model.rowCount():
+                # when there are transfer files found, we go to the configuration page for data import
                 return self.Page_ImportDataConfiguration_Id
             return self.Page_ProjectCreation_Id
 
         if self.current_id == self.Page_GenerateDatabaseSelection_Id:
-            # update configuration for project generation for import schema and use schema config to save
+            # update configuration for import data 
+            # and for import schema and export data (to have the same db settings for all of them)
+            # and use schema config to save
             self.generate_database_seletion_page.update_configuration(
                 self.import_schema_configuration)
             self.generate_database_seletion_page.update_configuration(
                 self.import_data_configuration)
+            self.generate_database_seletion_page.update_configuration(
+                self.export_data_configuration)
             self.generate_database_seletion_page.save_configuration(
                 self.import_schema_configuration)
             return self.Page_ProjectCreation_Id
+
+        if self.current_id == self.Page_ExportDatabaseSelection_Id:
+            # update configuration for export data 
+            # and for import schema and import data (to have the same db settings for all of them)
+            # and use schema config to save
+            self.export_database_seletion_page.update_configuration(
+                self.import_schema_configuration)
+            self.export_database_seletion_page.update_configuration(
+                self.import_data_configuration)
+            self.export_database_seletion_page.update_configuration(
+                self.export_data_configuration)
+            self.export_database_seletion_page.save_configuration(
+                self.import_schema_configuration)
+            return self.Page_ExportDataConfiguration_Id
 
         if self.current_id == self.Page_ImportSchemaConfiguration_Id:
             self.schema_configuration_page.update_configuration(
@@ -476,7 +231,8 @@ class WorkflowWizard (QWizard):
                 return self.Page_ImportSchemaExecution_Id
             if self.import_data_file_model.rowCount():
                 return self.Page_ImportDataConfiguration_Id
-            #dave else close
+            else:
+                self.log_panel.print_info(self.tr(f"No models, no transfer files, nothing to do..."))
 
         if self.current_id == self.Page_ImportSchemaExecution_Id:
             if self.import_data_file_model.rowCount():
@@ -488,19 +244,10 @@ class WorkflowWizard (QWizard):
 
         if self.current_id == self.Page_ImportDataExecution_Id:
             return self.Page_ProjectCreation_Id
-
-        if self.current_id == self.Page_ExportDatabaseSelection_Id:
-            # update configuration for import data and for import schema and use schema config to save
-            self.export_database_seletion_page.update_configuration(
-                self.export_data_configuration)
-            self.export_database_seletion_page.save_configuration(
-                self.export_data_configuration)
-            return self.Page_ExportDataConfiguration_Id
         
         if self.current_id == self.Page_ExportDataConfiguration_Id:
             return self.Page_ExportDataExecution_Id
         
-        # fallback
         return(self.current_id)
 
     def id_changed(self, new_id):
@@ -551,15 +298,53 @@ class WorkflowWizard (QWizard):
             sessions[self.current_export_target]['dataset']=self.export_datasets_model.selected_dataset()
             self.export_data_execution_page.setup_sessions(self.export_data_configuration, sessions)           
 
-    def refresh_import_models(self, silent=False):
-        db_connector=self._get_db_connector(self.import_schema_configuration, self.db_simple_factory)
-        return self.import_models_model.refresh_model(self.file_model, db_connector, silent)
+    def _current_page_title(self, id):
+        if id == self.Page_ImportSourceSeletion_Id:
+            return self.tr("Source Selection")
+        elif id == self.Page_ImportDatabaseSelection_Id:
+            return self.tr("Database Configuration")
+        elif id == self.Page_GenerateDatabaseSelection_Id:
+            return self.tr("Database Configuration")
+        elif id == self.Page_ImportSchemaConfiguration_Id:
+            return self.tr("Schema Import Configuration")
+        elif id == self.Page_ImportSchemaExecution_Id:
+            return self.tr("Schema Import Sessions")
+        elif id == self.Page_ImportDataConfiguration_Id:
+            return self.tr("Data import configuration")
+        elif id == self.Page_ImportDataExecution_Id:
+            return self.tr("Data Import Sessions")
+        elif id == self.Page_ExportDataConfiguration_Id:
+            return self.tr("Data export configuration")
+        elif id == self.Page_ExportDataExecution_Id:
+            return self.tr("Data export Sessions")
+        elif id == self.Page_ProjectCreation_Id:
+            return self.tr("Generate a QGIS Project")
+        else:
+            return self.tr("Model Baker - Workflow Wizard")
 
     def refresh_export_models(self):
-        db_connector=self._get_db_connector(self.export_data_configuration, self.db_simple_factory)
+        db_connector=self._get_db_connector(self.export_data_configuration)
         self.export_models_model.refresh_model(db_connector)
         self.export_datasets_model.refresh_model(db_connector)
         return 
+
+    def refresh_import_models(self, silent=False):
+        db_connector=self._get_db_connector(self.import_schema_configuration)
+        return self.import_models_model.refresh_model(self.source_model, db_connector, silent)
+
+    def _get_db_connector(self, configuration):
+        # migth be moved to db_utils...
+        db_simple_factory = DbSimpleFactory()
+        schema = configuration.dbschema
+
+        db_factory = db_simple_factory.create_factory(configuration.tool)
+        config_manager = db_factory.get_db_command_config_manager(configuration)
+        uri_string = config_manager.get_uri(configuration.db_use_super_login)
+
+        try:
+            return db_factory.get_db_connector(uri_string, schema)
+        except (DBConnectorError, FileNotFoundError):
+            return None
 
     def get_topping_file_list(self, id_list, log_panel):
         topping_file_model = self.get_topping_file_model(id_list, log_panel)
@@ -607,43 +392,6 @@ class WorkflowWizard (QWizard):
                 ' '.join(missing_file_ids)), LogPanel.COLOR_TOPPING)
 
         return topping_file_cache.model
-
-    def _current_page_title(self, id):
-        if id == self.Page_ImportSourceSeletion_Id:
-            return self.tr("Source Selection")
-        elif id == self.Page_ImportDatabaseSelection_Id:
-            return self.tr("Database Configuration")
-        elif id == self.Page_GenerateDatabaseSelection_Id:
-            return self.tr("Database Configuration")
-        elif id == self.Page_ImportSchemaConfiguration_Id:
-            return self.tr("Schema Import Configuration")
-        elif id == self.Page_ImportSchemaExecution_Id:
-            return self.tr("Schema Import Sessions")
-        elif id == self.Page_ImportDataConfiguration_Id:
-            return self.tr("Data import configuration")
-        elif id == self.Page_ImportDataExecution_Id:
-            return self.tr("Data Import Sessions")
-        elif id == self.Page_ExportDataConfiguration_Id:
-            return self.tr("Data export configuration")
-        elif id == self.Page_ExportDataExecution_Id:
-            return self.tr("Data export Sessions")
-        elif id == self.Page_ProjectCreation_Id:
-            return self.tr("Generate a QGIS Project")
-        else:
-            return self.tr("Model Baker - Workflow Wizard")
-
-    def _get_db_connector(self, configuration, db_simple_factory):
-        # migth be moved to db_utils...
-        schema = configuration.dbschema
-
-        db_factory = db_simple_factory.create_factory(configuration.tool)
-        config_manager = db_factory.get_db_command_config_manager(configuration)
-        uri_string = config_manager.get_uri(configuration.db_use_super_login)
-
-        try:
-            return db_factory.get_db_connector(uri_string, schema)
-        except (DBConnectorError, FileNotFoundError):
-            return None
 
 class WorkflowWizardDialog (QDialog):
